ai-web-browse/pyproject.toml
[project]
name = "ai-web-browse"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12",
    "litellm>=1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32",
    "rich>=13.0",
]

[project.scripts]
browse-website = "browse_web:main"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]
